Initial commit
Brings the core RAD models into this engine. This will ultimately
facilitate the sharing of logic between the acquisition and consumer
facing products.

Some tests were harmed in the making of this engine. Specifically those
that reached into the I18n stuff from the hosting application. The
things they covered are well specified in the RAD project so don't
sweat it.