Adds pagination for `Firm` ES queries
Moves creation of `SearchResult` to the `FirmRepository`. This is where
it should have been all along. The `SearchResult` is tied to `Firm` via
`FirmResult`, plus it is closer to the knowledge of current page, via
the repository.

Specifically supports the kaminari protocol, also providing metadata
to support the 'Showing X to X of X' style labelling.