Take ordering into account
When multiple sort filters are applied, the order of the returned sort
values becomes positional. In which case we should always return the
last entry as we will always sort by geographical distance last.

This is somewhat fragile at the moment, so we'll revisit if we apply
more sort filters.