Adds `friendly_name` attribute for advice methods
This relies on the existence of the correct I18n keys for types that
should exhibit this behaviour.

Relying on the ordinals for naming/translation is super-hacky and icky
feeling but will have to do for now at least. The rationale being the
same as elsewhere - that the ordinal is the only thing we can key off
across environments that we can be assured is the same.

It'll do.