Change business offering percentage fields to booleans.
I decided not to propagate this change to ElasticSearch, because that
would mean doing some mucking around with the live server to make all
the databases and codebases align. It's possible that we'll do that, but
given that this is a priority change and it'll be a while until this
is deployed to production — we'll defer it.

`firm_serializer` is responsible for converting booleans to percentages.