Add `Firm#remote_or_local_advice` for new remote/f2f question
This is an attribute that checks whether the firm has specified
in person or remote advice methods as their primary advice delivery
mechanism. Since we can infer this data from the advice methods selected
(and indeed, will need to since our current dataset doesn't have this
information and we can't ask for more at this stage) we do so.