Carry out the serializing and storing of a firm asynchronously
With very large firms indexing becomes slow.

A firm with 450 advisers takes around 3 seconds to completely index.
This causes a notable delay when clicking save in the self service web
UI.

Timing various parts of this process indicates that it is the HTTP PUT
to ElasticSearch which is taking the majority of this time.

This commit reinstates IndexFirmJob to do the indexing asynchronously.
So we can offload the delay from the web request.