Ask codeclimate to ignore snapshot model
Codeclimate does not like this class because it's "too complex". I'm
assuming that's because of the number of methods. The only reason to
move these methods to another class/module would be to satisfy
Codeclimate; I don't actually think the methods need to be moved. There
may be some way, through metaprogramming, to remove a lot of the
methods here.