Delete FirmResult#telephone_number wrapper
This method is obsolete. It should have been deleted as part of
https://github.com/moneyadviceservice/mas-rad_core/pull/146
But we missed it, assuming that telphone numbers only come from offices
now.

Essentially this was causing phone numbers for remote firms not to
appear as they are now pre-formatted in ElasticSearch and UKPhoneNumbers
doesn't cope with input numbers containing spaces and returns nil.

Instead we simply expose the field directly.