Only send the postcode when validating offices address
This will solve the main problem that Google is not
returning latitude and longitude for ambiguous address
that has building name on the address. This is a
limitation of the Geocoding API listed here:

https://maps-apis.googleblog.com/2016/11/address-geocoding-in-google-maps-apis.html

So the possible solutions were:

* Change to the Places API like suggested on the link above;
* Change to Bing
* Add a field name called Building name
* Just send the postcode to validate the address on Google.

After conversation with the stakeholders, they told to just send the postcode instead the address line 1 and address line 2.